+++ regression-test/suites/fault_injection_p0/flexible/test_f_auto_inc_compaction.groovy
@@ -20,10 +20,14 @@ import java.util.concurrent.TimeUnit
 import org.awaitility.Awaitility
 
 suite("test_f_auto_inc_compaction", "nonConcurrent") {
+    if (isCloudMode()) {
+        logger.info("skip test_f_auto_inc_compaction in cloud mode")
+        return
+    }
 
     GetDebugPoint().clearDebugPointsForAllFEs()
     GetDebugPoint().clearDebugPointsForAllBEs()
-
+    def dbName = context.config.getDbNameByFile(context.file)
     def tableName = "test_f_auto_inc_compaction"
     sql """ DROP TABLE IF EXISTS ${tableName} """
     sql """ CREATE TABLE ${tableName} (
@@ -96,130 +100,108 @@ suite("test_f_auto_inc_compaction", "nonConcurrent") {
         }
     })
 
-    def enable_publish_spin_wait = { tokenName -> 
-        if (isCloudMode()) {
-            GetDebugPoint().enableDebugPointForAllFEs("CloudGlobalTransactionMgr.getDeleteBitmapUpdateLock.enable_spin_wait", [token: "${tokenName}"])
-        } else {
-            GetDebugPoint().enableDebugPointForAllBEs("EnginePublishVersionTask::execute.tablet.enable_spin_wait", [token: "${tokenName}", tablet_id: "${tabletId}"])
-        }
+    def do_streamload_2pc_commit = { txnId ->
+        def command = "curl -X PUT --location-trusted -u ${context.config.feHttpUser}:${context.config.feHttpPassword}" +
+                " -H txn_id:${txnId}" +
+                " -H txn_operation:commit" +
+                " http://${context.config.feHttpAddress}/api/${dbName}/${tableName}/_stream_load_2pc"
+        log.info("http_stream execute 2pc: ${command}")
+
+        def process = command.execute()
+        def code = process.waitFor()
+        def out = process.text
+        def json2pc = parseJson(out)
+        log.info("http_stream 2pc result: ${out}".toString())
+        assertEquals(code, 0)
+        assertEquals("success", json2pc.status.toLowerCase())
     }
 
-    def disable_publish_spin_wait = {
-        if (isCloudMode()) {
-            GetDebugPoint().disableDebugPointForAllFEs("CloudGlobalTransactionMgr.getDeleteBitmapUpdateLock.enable_spin_wait")
-        } else {
-            GetDebugPoint().disableDebugPointForAllBEs("EnginePublishVersionTask::execute.tablet.enable_spin_wait")
+    def wait_for_publish = {txnId, waitSecond ->
+        String st = "PREPARE"
+        while (!st.equalsIgnoreCase("VISIBLE") && !st.equalsIgnoreCase("ABORTED") && waitSecond > 0) {
+            Thread.sleep(1000)
+            waitSecond -= 1
+            def result = sql_return_maparray "show transaction from ${dbName} where id = ${txnId}"
+            assertNotNull(result)
+            st = result[0].TransactionStatus
         }
-    }
-
-    def enable_block_in_publish = { passToken -> 
-        if (isCloudMode()) {
-            GetDebugPoint().enableDebugPointForAllFEs("CloudGlobalTransactionMgr.getDeleteBitmapUpdateLock.block", [pass_token: "${passToken}"])
-        } else {
-            GetDebugPoint().enableDebugPointForAllBEs("EnginePublishVersionTask::execute.tablet.block", [pass_token: "${passToken}", tablet_id: "${tabletId}"])
-        }
-    }
-
-    def disable_block_in_publish = {
-        if (isCloudMode()) {
-            GetDebugPoint().disableDebugPointForAllFEs("CloudGlobalTransactionMgr.getDeleteBitmapUpdateLock.block")
-        } else {
-            GetDebugPoint().disableDebugPointForAllBEs("EnginePublishVersionTask::execute.tablet.block")
-        }
+        log.info("Stream load with txn ${txnId} is ${st}")
+        assertEquals(st, "VISIBLE")
     }
 
 
-
-    try {
-        GetDebugPoint().clearDebugPointsForAllFEs()
-        GetDebugPoint().clearDebugPointsForAllBEs()
-
-
-        enable_publish_spin_wait("t1")
-        enable_block_in_publish("-1")
-
-        Thread.sleep(1000)
-        def t1 = Thread.start {
-            sql "insert into ${tableName} values(99,99,99,99);"
-        }
-        Thread.sleep(1000)
-
-
-        enable_publish_spin_wait("t2")
-        Thread.sleep(1000)
-        def t2 = Thread.start {
-            String load2 = 
-                    """{"k":3,"v1":100}
-                    {"k":6,"v2":600}"""
-            streamLoad {
-                table "${tableName}"
-                set 'format', 'json'
-                set 'read_json_by_line', 'true'
-                set 'unique_key_update_mode', 'UPDATE_FLEXIBLE_COLUMNS'
-                inputStream new ByteArrayInputStream(load2.getBytes())
-                time 60000
+    def txnId1, txnId2
+    String load1 = """99,99,99,99"""
+    streamLoad {
+        table "${tableName}"
+        set 'column_separator', ','
+        set 'format', 'csv'
+        set 'two_phase_commit', 'true'
+        inputStream new ByteArrayInputStream(load1.getBytes())
+        time 60000 // limit inflight 60s
+        check { result, exception, startTime, endTime ->
+            if (exception != null) {
+                throw exception
             }
+            log.info("Stream load result: ${result}".toString())
+            def json = parseJson(result)
+            txnId1 = json.TxnId
+            assertEquals("success", json.Status.toLowerCase())
         }
-        Thread.sleep(1000)
-
-
-        // most have a new load after load2's max version sees in flush phase. Otherwise load2 will skip to calc delete bitmap
-        // of rowsets which is produced by compaction and whose end version is lower than the max version
-        // the load1 sees in flush phase, see https://github.com/apache/doris/pull/38487 for details
-
-        // let load1 publish
-        enable_block_in_publish("t1")
-        t1.join()
-        Thread.sleep(1000)
-
-
-        // trigger full compaction on tablet
-        logger.info("trigger compaction on another BE ${tabletBackend.Host} with backendId=${tabletBackend.BackendId}")
-        def (code, out, err) = be_run_full_compaction(tabletBackend.Host, tabletBackend.HttpPort, tabletId)
-        logger.info("Run compaction: code=" + code + ", out=" + out + ", err=" + err)
-        Assert.assertEquals(code, 0)
-        def compactJson = parseJson(out.trim())
-        Assert.assertEquals("success", compactJson.status.toLowerCase())
-
-        // wait for full compaction to complete
-        Awaitility.await().atMost(3, TimeUnit.SECONDS).pollDelay(200, TimeUnit.MILLISECONDS).pollInterval(100, TimeUnit.MILLISECONDS).until(
-            {
-                (code, out, err) = be_get_compaction_status(tabletBackend.Host, tabletBackend.HttpPort, tabletId)
-                logger.info("Get compaction status: code=" + code + ", out=" + out + ", err=" + err)
-                Assert.assertEquals(code, 0)
-                def compactionStatus = parseJson(out.trim())
-                Assert.assertEquals("success", compactionStatus.status.toLowerCase())
-                return !compactionStatus.run_status
-            }
-        )
-
-        check_rs_metas(1, {int startVersion, int endVersion, int numSegments, int numRows, String overlapPb ->
-            // check the rowset produced by full compaction
-            // [0-6]
-            Assert.assertEquals(startVersion, 0)
-            Assert.assertEquals(endVersion, 6)
-            Assert.assertEquals(numRows, 4)
-            Assert.assertEquals(overlapPb, "NONOVERLAPPING")
-        })
-
-        qt_after_compaction "select k,v1,v2 from ${tableName} order by k;"
-        qt_check_auto_inc_dup "select count(*) from ${tableName} group by id having count(*) > 1;"
-
-        disable_publish_spin_wait()
-        disable_block_in_publish()
-        Thread.sleep(1000)
-        t2.join()
-
-        qt_sql "select k,v1,v2 from ${tableName} order by k;"
-        qt_check_auto_inc_dup "select count(*) from ${tableName} group by id having count(*) > 1;"
-        qt_check_auto_inc_val "select count(*) from ${tableName} where id=0;"
-    } catch(Exception e) {
-        logger.info(e.getMessage())
-        throw e
-    } finally {
-        disable_publish_spin_wait()
-        disable_block_in_publish()
-        GetDebugPoint().clearDebugPointsForAllFEs()
-        GetDebugPoint().clearDebugPointsForAllBEs()
     }
+
+
+    String load2 = 
+            """{"k":3,"v1":100}
+            {"k":6,"v2":600}"""
+    streamLoad {
+        table "${tableName}"
+        set 'format', 'json'
+        set 'read_json_by_line', 'true'
+        set 'two_phase_commit', 'true'
+        set 'unique_key_update_mode', 'UPDATE_FLEXIBLE_COLUMNS'
+        inputStream new ByteArrayInputStream(load2.getBytes())
+        time 60000
+        check { result, exception, startTime, endTime ->
+            if (exception != null) {
+                throw exception
+            }
+            log.info("Stream load result: ${result}".toString())
+            def json = parseJson(result)
+            txnId2 = json.TxnId
+            assert "success" == json.Status.toLowerCase()
+        }
+    }
+
+
+    // most have a new load after load2's max version sees in flush phase. Otherwise load2 will skip to calc delete bitmap
+    // of rowsets which is produced by compaction and whose end version is lower than the max version
+    // the load1 sees in flush phase, see https://github.com/apache/doris/pull/38487 for details
+
+    // let load1 publish
+    do_streamload_2pc_commit(txnId1)
+    wait_for_publish(txnId1, 10)
+    Thread.sleep(1000)
+
+
+    trigger_and_wait_compaction(tableName, "full")
+
+    check_rs_metas(1, {int startVersion, int endVersion, int numSegments, int numRows, String overlapPb ->
+        // check the rowset produced by full compaction
+        // [0-6]
+        Assert.assertEquals(startVersion, 0)
+        Assert.assertEquals(endVersion, 6)
+        Assert.assertEquals(numRows, 4)
+        Assert.assertEquals(overlapPb, "NONOVERLAPPING")
+    })
+
+    qt_after_compaction "select k,v1,v2 from ${tableName} order by k;"
+    qt_check_auto_inc_dup "select count(*) from ${tableName} group by id having count(*) > 1;"
+
+    do_streamload_2pc_commit(txnId2)
+    wait_for_publish(txnId2, 10)
+
+    qt_sql "select k,v1,v2 from ${tableName} order by k;"
+    qt_check_auto_inc_dup "select count(*) from ${tableName} group by id having count(*) > 1;"
+    qt_check_auto_inc_val "select count(*) from ${tableName} where id=0;"
 }
