+++ regression-test/suites/fault_injection_p0/flexible/test_f_seq_publish_read_from_old.groovy
@@ -18,10 +18,15 @@
 import org.junit.Assert
 
 suite("test_f_seq_publish_read_from_old", "nonConcurrent") {
+    if (isCloudMode()) {
+        logger.info("skip test_f_seq_publish_read_from_old in cloud mode")
+        return
+    }
 
     GetDebugPoint().clearDebugPointsForAllFEs()
     GetDebugPoint().clearDebugPointsForAllBEs()
 
+    def dbName = context.config.getDbNameByFile(context.file)
     def tableName = "test_f_seq_publish_read_from_old"
     sql """ DROP TABLE IF EXISTS ${tableName} """
     sql """ CREATE TABLE ${tableName} (
@@ -41,10 +46,6 @@ suite("test_f_seq_publish_read_from_old", "nonConcurrent") {
         "store_row_column" = "false"); """
     sql """insert into ${tableName} values(1,100,1,1,1,1),(2,100,2,2,2,2),(3,100,3,3,3,3),(4,100,4,4,4,4);"""
     qt_sql "select k,v1,v2,v3,v4,v5 from ${tableName} order by k;"
-
-    def tabletStat = sql_return_maparray("show tablets from ${tableName};").get(0)
-    def tabletBackendId = tabletStat.BackendId
-    def tabletId = tabletStat.TabletId
     
     def inspectRows = { sqlStr ->
         sql "set skip_delete_sign=true;"
@@ -56,92 +57,90 @@ suite("test_f_seq_publish_read_from_old", "nonConcurrent") {
         sql "sync"
     }
 
-    def enable_publish_spin_wait = {
-        if (isCloudMode()) {
-            GetDebugPoint().enableDebugPointForAllFEs("CloudGlobalTransactionMgr.getDeleteBitmapUpdateLock.enable_spin_wait")
-        } else {
-            GetDebugPoint().enableDebugPointForAllBEs("EnginePublishVersionTask::execute.tablet.enable_spin_wait", [tablet_id: "${tabletId}"])
-        }
+    def do_streamload_2pc_commit = { txnId ->
+        def command = "curl -X PUT --location-trusted -u ${context.config.feHttpUser}:${context.config.feHttpPassword}" +
+                " -H txn_id:${txnId}" +
+                " -H txn_operation:commit" +
+                " http://${context.config.feHttpAddress}/api/${dbName}/${tableName}/_stream_load_2pc"
+        log.info("http_stream execute 2pc: ${command}")
+
+        def process = command.execute()
+        def code = process.waitFor()
+        def out = process.text
+        def json2pc = parseJson(out)
+        log.info("http_stream 2pc result: ${out}".toString())
+        assertEquals(code, 0)
+        assertEquals("success", json2pc.status.toLowerCase())
     }
 
-    def disable_publish_spin_wait = {
-        if (isCloudMode()) {
-            GetDebugPoint().disableDebugPointForAllFEs("CloudGlobalTransactionMgr.getDeleteBitmapUpdateLock.enable_spin_wait")
-        } else {
-            GetDebugPoint().disableDebugPointForAllBEs("EnginePublishVersionTask::execute.tablet.enable_spin_wait")
+    def wait_for_publish = {txnId, waitSecond ->
+        String st = "PREPARE"
+        while (!st.equalsIgnoreCase("VISIBLE") && !st.equalsIgnoreCase("ABORTED") && waitSecond > 0) {
+            Thread.sleep(1000)
+            waitSecond -= 1
+            def result = sql_return_maparray "show transaction from ${dbName} where id = ${txnId}"
+            assertNotNull(result)
+            st = result[0].TransactionStatus
         }
+        log.info("Stream load with txn ${txnId} is ${st}")
+        assertEquals(st, "VISIBLE")
     }
 
-    def enable_block_in_publish = {
-        if (isCloudMode()) {
-            GetDebugPoint().enableDebugPointForAllFEs("CloudGlobalTransactionMgr.getDeleteBitmapUpdateLock.block")
-        } else {
-            GetDebugPoint().enableDebugPointForAllBEs("EnginePublishVersionTask::execute.tablet.block", [tablet_id: "${tabletId}"])
-        }
-    }
-
-    def disable_block_in_publish = {
-        if (isCloudMode()) {
-            GetDebugPoint().disableDebugPointForAllFEs("CloudGlobalTransactionMgr.getDeleteBitmapUpdateLock.block")
-        } else {
-            GetDebugPoint().disableDebugPointForAllBEs("EnginePublishVersionTask::execute.tablet.block")
-        }
-    }
-
-    try {
-        enable_publish_spin_wait()
-        enable_block_in_publish()
-
-        def t1 = Thread.start {
-            String load1 = 
-                """{"k":1,"v1":100,"__DORIS_DELETE_SIGN__":1}
-                   {"k":2,"v1":200,"__DORIS_DELETE_SIGN__":1}
-                   {"k":3,"v1":50,"__DORIS_DELETE_SIGN__":1}
-                   {"k":4,"__DORIS_DELETE_SIGN__":1}"""
-            streamLoad {
-                table "${tableName}"
-                set 'format', 'json'
-                set 'read_json_by_line', 'true'
-                set 'unique_key_update_mode', 'UPDATE_FLEXIBLE_COLUMNS'
-                inputStream new ByteArrayInputStream(load1.getBytes())
-                time 60000
+    def txnId1, txnId2
+    String load1 = 
+        """{"k":1,"v1":100,"__DORIS_DELETE_SIGN__":1}
+            {"k":2,"v1":200,"__DORIS_DELETE_SIGN__":1}
+            {"k":3,"v1":50,"__DORIS_DELETE_SIGN__":1}
+            {"k":4,"__DORIS_DELETE_SIGN__":1}"""
+    streamLoad {
+        table "${tableName}"
+        set 'format', 'json'
+        set 'read_json_by_line', 'true'
+        set 'two_phase_commit', 'true'
+        set 'unique_key_update_mode', 'UPDATE_FLEXIBLE_COLUMNS'
+        inputStream new ByteArrayInputStream(load1.getBytes())
+        time 60000
+        check { result, exception, startTime, endTime ->
+            if (exception != null) {
+                throw exception
             }
+            log.info("Stream load result: ${result}".toString())
+            def json = parseJson(result)
+            txnId1 = json.TxnId
+            assert "success" == json.Status.toLowerCase()
         }
-
-        Thread.sleep(1500)
-        def t2 = Thread.start {
-            String load2 = 
-                """{"k":1,"v2":987,"v3":77777}
-                   {"k":2,"v2":987,"v3":77777}
-                   {"k":3,"v2":987,"v3":77777}
-                   {"k":4,"v2":987,"v3":77777}"""
-            streamLoad {
-                table "${tableName}"
-                set 'format', 'json'
-                set 'read_json_by_line', 'true'
-                set 'unique_key_update_mode', 'UPDATE_FLEXIBLE_COLUMNS'
-                inputStream new ByteArrayInputStream(load2.getBytes())
-                time 60000
-            }
-        }
-
-        Thread.sleep(2000)
-        disable_block_in_publish()
-        t1.join()
-        t2.join()
-
-        Thread.sleep(1500)
-        sql "sync;"
-        qt_sql "select k,v1,v2,v3,v4,v5 from ${tableName} order by k;"
-        inspectRows "select k,v1,v2,v3,v4,v5,__DORIS_SEQUENCE_COL__,__DORIS_VERSION_COL__,BITMAP_TO_STRING(__DORIS_SKIP_BITMAP_COL__),__DORIS_DELETE_SIGN__ from ${tableName} order by k,__DORIS_VERSION_COL__,__DORIS_SEQUENCE_COL__;"
-
-    } catch(Exception e) {
-        logger.info(e.getMessage())
-        throw e
-    } finally {
-        disable_publish_spin_wait()
-        disable_block_in_publish()
-        GetDebugPoint().clearDebugPointsForAllFEs()
-        GetDebugPoint().clearDebugPointsForAllBEs()
     }
+
+    String load2 = 
+        """{"k":1,"v2":987,"v3":77777}
+            {"k":2,"v2":987,"v3":77777}
+            {"k":3,"v2":987,"v3":77777}
+            {"k":4,"v2":987,"v3":77777}"""
+    streamLoad {
+        table "${tableName}"
+        set 'format', 'json'
+        set 'read_json_by_line', 'true'
+        set 'two_phase_commit', 'true'
+        set 'unique_key_update_mode', 'UPDATE_FLEXIBLE_COLUMNS'
+        inputStream new ByteArrayInputStream(load2.getBytes())
+        time 60000
+        check { result, exception, startTime, endTime ->
+            if (exception != null) {
+                throw exception
+            }
+            log.info("Stream load result: ${result}".toString())
+            def json = parseJson(result)
+            txnId2 = json.TxnId
+            assert "success" == json.Status.toLowerCase()
+        }
+    }
+    do_streamload_2pc_commit(txnId1)
+    wait_for_publish(txnId1, 10)
+
+    do_streamload_2pc_commit(txnId2)
+    wait_for_publish(txnId2, 10)
+
+    sql "sync;"
+    qt_sql "select k,v1,v2,v3,v4,v5 from ${tableName} order by k;"
+    inspectRows "select k,v1,v2,v3,v4,v5,__DORIS_SEQUENCE_COL__,__DORIS_VERSION_COL__,BITMAP_TO_STRING(__DORIS_SKIP_BITMAP_COL__),__DORIS_DELETE_SIGN__ from ${tableName} order by k,__DORIS_VERSION_COL__,__DORIS_SEQUENCE_COL__;"
 }
